BLE updated and tested

+++ Drivers/Inc/Timer.h
@@ -136,7 +136,7 @@ DEFINES
 /* Timer4 CCER OC4P - Capture/Compare 4 Output Polarity offset */
 #define TIM4_CCER_OC4P_OFFSET  (13UL)
 
-/* Timer4 CNT - Timer4 Counter Register */
+/* Timer4 CNT - Timer4 Counter Register - Used mainly for debugging purposes */
 #define TIM4_CNT_OFFSET        (0x24UL)
 #define TIM4_CNT_BASE_ADDRESS  (TIM4_BASE_ADDRESS + TIM4_CNT_OFFSET) // typecast and dereference
 #define TIM4_CNT          (* (volatile uint32_t *)(TIM4_CNT_BASE_ADDRESS)) // typecast and dereference
@@ -180,7 +180,7 @@ FUNCTIONS PROTOTYPES
 ****************************************************************************/
 
 /* Create delay in milliseconds */
-void Delay_ms(uint32_t ms);
+void Delay_ms(float ms);
 
 /* Initialize CPU SysTick and General-Purpose Timer peripheral */
 void Timer_Init();
