+++ Drivers/Src/Timer.c
@@ -12,10 +12,10 @@ extern volatile uint32_t SysTick_Counter;
 
 
 
-void Delay_ms(uint32_t ms)
+void Delay_ms(float ms)
 {
-	uint8_t Timer_Delay_Start = SysTick_Counter;
-	while((SysTick_Counter - Timer_Delay_Start) < ms); // Wait inside this function until the expected number of ms are elapsed
+	uint32_t Timer_Delay_Start = SysTick_Counter;
+	while((float)(SysTick_Counter - Timer_Delay_Start) < ms); // Wait inside this function until the expected number of ms are elapsed
 }
 
 
@@ -30,8 +30,8 @@ void Timer_Init(void)
 	SYSTICK_CTRL |= (0x01UL << SYSTICK_CTRL_ENABLE_OFFSET); // Enable counter
 	/* SysTick Exception Request Enable */
 	SYSTICK_CTRL |= (0x01UL << SYSTICK_CTRL_TICKINT_OFFSET); // Assert the SysTick exception request
-
-	SYSTICK_CTRL |= (0x01UL << SYSTICK_CTRL_CLKSOURCE_OFFSET);
+	/* Processor Clock selection */
+	SYSTICK_CTRL |= (0x01UL << SYSTICK_CTRL_CLKSOURCE_OFFSET); // Processor Clock (AHB) selected
 
 	/* Timer4 peripheral initialization */
 	/* Enable Clock for Timer4 peripheral */
